move Node.followNextEOG to correct file
